+++ packages/script/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @baustatik/script
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [6c215da]
+- Updated dependencies [6c215da]
+- Updated dependencies [6c215da]
+  - @baustatik/cross-section@0.0.9
+
 ## 0.0.9
 
 ### Patch Changes